gpu shared types autogen refactor
for webgpu
