+++ crates/uzu/src/backends/metal/kernel/embedding/quant_embedding.metal
@@ -21,6 +21,9 @@ PUBLIC KERNEL(QuantizedEmbeddingLookup) (
     const uint dim_idx AXIS(model_dim, 16),
     const uint batch_idx AXIS(batch_size, 16)
 ) {
+  const QuantizationMode quantization_mode = QuantizationMode(
+      quant_mode
+  ); // TODO: should be accepted as a kernel argument
   const uint thread_position_in_grid = batch_idx * model_dim + dim_idx;
   const uint64_t token_id = token_ids[batch_idx];
   if (token_id >= vocab_size) {
@@ -34,12 +37,13 @@ PUBLIC KERNEL(QuantizedEmbeddingLookup) (
   const T scale = scales[token_id * num_groups + group_idx];
   const T bias = biases[token_id * num_groups + group_idx];
 
-  const uint packing_divisor = quant_mode == QuantizationMode::UINT4 ? 2 : 1;
+  const uint packing_divisor =
+      quantization_mode == QuantizationMode::UINT4 ? 2 : 1;
   const uint weights_stride = model_dim / packing_divisor;
 
   int quantized_value = 0;
-  switch (quant_mode) {
-  case UINT4: {
+  switch (quantization_mode) {
+  case QuantizationMode::UINT4: {
     const uint byte_idx = token_id * weights_stride + (dim_idx / 2);
     const uint8_t packed = weights[byte_idx];
     if ((dim_idx & 1) == 0) {
@@ -49,14 +53,14 @@ PUBLIC KERNEL(QuantizedEmbeddingLookup) (
     }
     break;
   };
-  case INT8: {
+  case QuantizationMode::INT8: {
     const uint elem_idx = token_id * weights_stride + dim_idx;
     const device int8_t* weights_i8 =
         reinterpret_cast<const device int8_t*>(weights);
     quantized_value = int(weights_i8[elem_idx]);
     break;
   };
-  case UINT8: {
+  case QuantizationMode::UINT8: {
     const uint elem_idx = token_id * weights_stride + dim_idx;
     quantized_value = int(weights[elem_idx]);
     break;
