+++ crates/uzu/src/backends/metal/kernel/matmul/common/gemm.h
@@ -2,7 +2,7 @@
 
 #include "loader.h"
 #include "threadgroup_tile.h"
-#include "../../../generated/matmul.h"
+#include "../../generated/matmul.h"
 
 using namespace metal;
