+++ crates/uzu/src/backends/metal/kernel/matmul/common/gemm_mpp_core.h
@@ -2,7 +2,7 @@
 
 #include "defines.h"
 #include "loader.h"
-#include "../../../generated/matmul.h"
+#include "../../generated/matmul.h"
 
 #include <MetalPerformancePrimitives/MetalPerformancePrimitives.h>
